test: extract goose simulator logic to a utility file
